# Perf/fast unions