@jeswr 2022-08-16
> In the case of sources being lazily generated in an asynchronous manner, ensuring correct error propagation and disposal of sources would be best addressed in a dedicated sources iterator that is bound to be context-specific.

So then I guess the only remaining question for this PR is when an array of sources is passed into the `UnionIterator` whether we should immediately start listening for the error event by subscribing that listener in the constructor.