@jacoscaz 2022-08-16
> So then I guess the only remaining question for this PR is when an array of sources is passed into the `UnionIterator` whether we should immediately start listening for the error event by subscribing that listener in the constructor.

I don't think so as that would lead to slightly different behaviors in the case of sources passed as an array vs. sources passed as an iterator. Wrapping the array into an iterator with `wrap()` and consuming it as if it had always been an iterator leads to more consistent behavior.