@RubenVerborgh 2022-08-22
@jeswr Some questions about the circular list:
1. Shall we extract `CircularList` to a separate file like `LinkedList`?
2. Is a `CircularList` strictly needed? I.e., do the performance gains really come from not having an array for the iterators?

I do agree that, conceptually, it is indeed a circular list though.

Given that the number of iterators is assumed to be small (10–20ish?) and not often changed (?), I am curious if there is a big impact.

> reasons for a major version bump

So this is a v4? (I followed the links, but no result for keyword _major_.)