@jeswr 2022-08-22
> Shall we extract `CircularList` to a separate file like `LinkedList`?

That's probably sensible. When I was first writing I didn't have clarity on what APIs would need to be exposed; but I can now try and refactor that retrospectively. 

> Is a CircularList strictly needed? I.e., do the performance gains really come from not having an array for the iterators?

The reason I have introduced it is that we need to always be able to delete an element in `O(1)` time when the `end` event for that particular iterator is called; which we cannot do with an array or the current `LinkedList` as both require the traversal of `O(n)` elements (where `n` is index of the element in the array/linkedlist). 

> So this is a v4? (I followed the links, but no result for keyword major.)

Yes, since the erroring behavior will change in some cases this will need to be a v4.