@RubenVerborgh 2022-08-22
> The reason I have introduced it is that we need to always be able to delete an element in `O(1)` time when the `end` event for that particular iterator is called;

I agree; but it seems to be that it would be `O(10)` at most; unless we are looking at scenarios where the number of iterators is really large?

I guess I'm simultaneously asking: what are the main sources of performance improvement in this PR? Is it indeed this new list structure?