@jeswr 2022-08-22
> Maybe OrderedSet extends LinkedList, adding O(1) has and delete operations via a Set or Map?

I'll have a crack at this in the morning (jet lag is starting to hit me hard). To confirm - this would be *preferred* over a circular linked list implementation (provided this has similar performance to the current implementation, which it should)