@RubenVerborgh 2022-08-22
> this would be _preferred_ over a circular linked list implementation

Not necessarily; my remark was more that all we need is O(1) deletion.

Circular list as such doesn't offer that; it's first an O(n) lookup, and only then deletion is O(1).
Which is why the Symbol lookup trick was added.

So what I'm arguing is that the Symbol lookup trick is basically what we need (and that we can also implement it with a Map); and that we get O(1) deletion by making it point at the _previous_ node rather than current node.