@jeswr 2022-08-22
> Could you details these a bit?

See https://github.com/RubenVerborgh/AsyncIterator/pull/79#discussion_r934929002. The reason that test broke is because we do not subscribe to the error event of iterators until after we start reading from them. So if the iterator errors before we first call `.read()` on it, then the error will not be forwarded.